add play time
The play time variable in the `.run()` method is here to define how many time a player can play in a row during this turn.

For example, if it's set to 3, the player can play his 3 characters before letting the opponent play his turn.

If it's set to 1, the player can play only one character before letting the opponent play.